refactor(web): consolidate form-designer into form-builder tool
Remove the linear form-builder tool; rename form-designer to take over
the form-builder id/route/display ("Form Builder / 表單建構器"). The 2D
canvas tool is now the sole form tool under id `form-builder`.

- git rm form-builder/{index,messages,ui}.tsx (linear tool gone)
- git mv form-designer/* into form-builder/ (all files + inspector/)
- Fix import in ui.tsx: ../form-builder/sample → ./sample
- Rename FormDesignerTool → FormBuilderTool in ui.tsx + ui.spec.tsx
- Update messages.ts: key form-designer→form-builder, titles updated
- web-core/tools.ts: remove linear entry; rename form-designer→form-builder
- web-core/packages.ts: relatedTools drop form-designer
- tools/index.ts + messages.ts: drop formDesigner import/entry
- index.spec.ts: remove "form-designer" from EXPECTED_WEB_TOOL_IDS
- Clean up stale comments referencing form-designer
