+++ apps/web/src/tools/form-builder/inspector/misc-sections.spec.tsx
