+++ apps/web/src/tools/form-builder/inspector/settings-panel.spec.tsx
