--- apps/web/src/tools/form-designer/index.ts
@@ -1,5 +1,0 @@
-import type { ToolModule } from "@/tools/types";
-
-import { FormDesignerTool } from "./ui";
-
-export const tool: ToolModule = { id: "form-designer", Component: FormDesignerTool };
