--- apps/web/src/tools/form-designer/messages.ts
@@ -1,21 +1,0 @@
-import type { LocaleMessages } from "@/tools/types";
-
-export const messages: LocaleMessages = {
-  en: {
-    Tools: {
-      "form-designer": {
-        title: "Form Designer",
-        description:
-          "A 2D form designer — drag fields on a 12-column grid, group them, set per-field config, and preview live.",
-      },
-    },
-  },
-  "zh-TW": {
-    Tools: {
-      "form-designer": {
-        title: "表單設計器",
-        description: "2D 表單設計器：在 12 欄網格上拖放欄位、分組、設定每欄屬性,並即時預覽。",
-      },
-    },
-  },
-};
