+++ apps/web/src/tools/index.ts
@@ -5,7 +5,6 @@ import type { ToolModule } from "./types";
 import { tool as dataFilterBuilder } from "./data-filter-builder";
 import { tool as dataFilterTester } from "./data-filter-tester";
 import { tool as formBuilder } from "./form-builder";
-import { tool as formDesigner } from "./form-designer";
 import { tool as bpmnViewer } from "./bpmn-viewer";
 import { tool as esClientDemo } from "./es-client-demo";
 import { tool as esQueryBuilder } from "./es-query-builder";
@@ -35,7 +34,6 @@ export const toolModules: ToolModule[] = [
   esClientDemo,
   pgFilterBuilder,
   formBuilder,
-  formDesigner,
   bpmnViewer,
 ];
