+++ apps/web/src/tools/messages.ts
@@ -3,7 +3,6 @@ import type { LocaleMessages } from "./types";
 import { messages as dataFilterBuilder } from "./data-filter-builder/messages";
 import { messages as dataFilterTester } from "./data-filter-tester/messages";
 import { messages as formBuilder } from "./form-builder/messages";
-import { messages as formDesigner } from "./form-designer/messages";
 import { messages as bpmnViewer } from "./bpmn-viewer/messages";
 import { messages as esClientDemo } from "./es-client-demo/messages";
 import { messages as esQueryBuilder } from "./es-query-builder/messages";
@@ -33,6 +32,5 @@ export const toolMessages: LocaleMessages[] = [
   esClientDemo,
   pgFilterBuilder,
   formBuilder,
-  formDesigner,
   bpmnViewer,
 ];
